pyproject.toml
[project]
name = "seasonalgeo"
version = "0.1.0"
description = "Multi-temporal satellite imagery retrieval for season-invariant UAV geo-localization"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "Pillow>=10.0",
    "tqdm>=4.65",
    "rich>=13.0",
    "pyyaml>=6.0",
    "requests>=2.28",
    "numpy>=1.24",
]

[project.optional-dependencies]
gee = [
    "earthengine-api>=1.4",
    "geemap>=0.34",
    "rasterio>=1.3",
]
planet = [
    "planet>=2.0",
    "rasterio>=1.3",
]
dev = ["pytest>=7.0", "ruff"]

[project.scripts]
seasonalgeo-parse = "seasonalgeo.scripts.s01_parse_datasets:main"
seasonalgeo-retrieve = "seasonalgeo.scripts.s02_retrieve_imagery:main"
seasonalgeo-format = "seasonalgeo.scripts.s03_format_dataset:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["seasonalgeo*"]

[tool.setuptools.package-data]
seasonalgeo = ["configs/*.yaml"]
